--story=1003107 --user=林邦泉 条件分流节点异常提示 https://www.tapd.cn/54247054/s/1076322

+++ src/main/java/codedriver/module/process/condition/handler/ProcessTaskStartTimeCondition.java
@@ -71,7 +71,7 @@ public class ProcessTaskStartTimeCondition extends ProcessTaskConditionBase impl
 				List<String> valueList = JSON.parseArray(JSON.toJSONString(value), String.class);
 				List<String> textList = new ArrayList<>();
 				for(String valueStr : valueList) {
-					textList.add(simpleDateFormat.format(new Date(Integer.parseInt(valueStr))));
+					textList.add(simpleDateFormat.format(new Date(Long.parseLong(valueStr))));
 				}
 				return textList;
 			}
